Refactor environment setup and share models
