+++ README.md
@@ -1,27 +1,27 @@
-# WebAngular
+# Angular Blog
 
-This project was generated with [Angular CLI](https://github.com/angular/angular-cli) version 16.0.0.
+This project is a simple blog built with Angular. It includes a blog list, article detail view and authentication pages.
 
-## Development server
+## Requirements
+- Node.js 16+
+- [Angular CLI](https://angular.io/cli)
 
-Run `ng serve` for a dev server. Navigate to `http://localhost:4200/`. The application will automatically reload if you change any of the source files.
+## Setup
+1. Install dependencies:
+   ```bash
+   npm install
+   ```
+2. Configure the backend URL in `src/environments/environment.ts` and `src/environments/environment.prod.ts`.
+3. Run the development server:
+   ```bash
+   ng serve
+   ```
+   The application will be available at `http://localhost:4200/`.
 
-## Code scaffolding
+## Testing
+The project uses Karma for unit tests. Execute:
+```bash
+npm test
+```
 
-Run `ng generate component component-name` to generate a new component. You can also use `ng generate directive|pipe|service|class|guard|interface|enum|module`.
-
-## Build
-
-Run `ng build` to build the project. The build artifacts will be stored in the `dist/` directory.
-
-## Running unit tests
-
-Run `ng test` to execute the unit tests via [Karma](https://karma-runner.github.io).
-
-## Running end-to-end tests
-
-Run `ng e2e` to execute the end-to-end tests via a platform of your choice. To use this command, you need to first add a package that implements end-to-end testing capabilities.
-
-## Further help
-
-To get more help on the Angular CLI use `ng help` or go check out the [Angular CLI Overview and Command Reference](https://angular.io/cli) page.
+Make sure Angular CLI is installed globally if the `ng` command is missing.
