+++ src/environments/environment.prod.ts
@@ -1,0 +1,4 @@
+export const environment = {
+  production: true,
+  backendUrl: 'https://api.example.com',
+};
